renamed models -> utils.py since that's the django convention
