+++ fennec/apps/diagram/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
-from apps.diagram.models import Index, ForeignKey, Schema, Layer, TableElement, RelationshipElement, Diagram
-from apps.diagram.models import Table, Namespace, Column
+from apps.diagram.utils import Index, ForeignKey, Schema, Layer, TableElement, RelationshipElement, Diagram
+from apps.diagram.utils import Table, Namespace, Column
 
 
 # def serialize(obj):
