+++ fennec/apps/diagram/utils.py
