+++ fennec/apps/versioncontroll/utils.py
@@ -2,7 +2,7 @@ from StringIO import StringIO
 
 from rest_framework.parsers import JSONParser
 
-from apps.diagram.models import ProjectBasicInfo, BranchBasicInfo, SandboxBasicInfo
+from apps.diagram.utils import ProjectBasicInfo, BranchBasicInfo, SandboxBasicInfo
 from fennec.apps.diagram.serializers import SchemaSerializer, NamespaceSerializer, TableSerializer, ColumnSerializer, \
     IndexSerializer, ForeignKeySerializer, LayerSerializer, TableElementSerializer, RelationshipElementSerializer, \
     DiagramSerializer
